Factored out base class for SQLInsertClause, SQLUpdateClause, SQLDeleteClause
